+++ src/components/Auth/ContentArea.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Area = styled.div`
+  margin-top: 60px;
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+`;
+
+export default Area;
