+++ src/components/Auth/TitleWrapper.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const TitleWrapper = styled.div`
+  max-width: '10px';
+  margin: auto auto;
+  -ms-user-select: none;
+  -moz-user-select: -moz-none;
+  -webkit-user-select: none;
+  -khtml-user-select: none;
+  user-select: none;
+`;
+
+export default TitleWrapper;
