+++ src/components/InnerAPI/AfterSignupHandler.js
@@ -1,0 +1,5 @@
+const onSignUpTask = async () => {
+  alert('signup');
+};
+
+export default onSignUpTask;
